+++ src/c++/uds/src/uds/hash-utils.h
@@ -42,20 +42,18 @@ static inline u32 extract_sampling_bytes(const struct uds_record_name *name)
 }
 
 /* Compute the chapter delta list for a given name. */
-static inline unsigned int
+static inline u32
 hash_to_chapter_delta_list(const struct uds_record_name *name, const struct geometry *geometry)
 {
-	return (unsigned int) ((extract_chapter_index_bytes(name) >>
-				geometry->chapter_address_bits) &
-			       ((1 << geometry->chapter_delta_list_bits) - 1));
+	return ((extract_chapter_index_bytes(name) >> geometry->chapter_address_bits) &
+		((1 << geometry->chapter_delta_list_bits) - 1));
 }
 
 /* Compute the chapter delta address for a given name. */
-static inline unsigned int
+static inline u32
 hash_to_chapter_delta_address(const struct uds_record_name *name, const struct geometry *geometry)
 {
-	return (unsigned int) (extract_chapter_index_bytes(name) &
-			       ((1 << geometry->chapter_address_bits) - 1));
+	return extract_chapter_index_bytes(name) & ((1 << geometry->chapter_address_bits) - 1);
 }
 
 static inline unsigned int
